+++ FrontEnd/package.json
@@ -26,7 +26,8 @@
     "react-native-unimodules": "~0.10.1",
     "react-native-web": "~0.11.7",
     "react-redux": "^7.2.1",
-    "redux": "^4.0.5"
+    "redux": "^4.0.5",
+    "use-force-update": "^1.0.7"
   },
   "devDependencies": {
     "@babel/core": "~7.9.0",
@@ -34,6 +35,7 @@
     "expo-cli": "^3.27.8",
     "jest": "~25.2.6",
     "react-test-renderer": "~16.11.0",
+    "redux-devtools-extension": "^2.13.8",
     "rimraf": "^3.0.2"
   },
   "jest": {
