--- FrontEnd/src/reducers/testreducer.js
@@ -1,6 +1,0 @@
-export default (state, action) => {
-    switch(action.type){
-        default:
-            return state || null;
-    }
-}
