+++ FrontEnd/src/views/Timetable.js
@@ -1,0 +1,10 @@
+import React from "react";
+import Stundenplan from "../images/Stundenplan_Mock.jpg"
+
+export default function Timetable() {
+    return (
+        <div>
+            <img src={Stundenplan} alt="this is car image" />
+        </div>
+    )
+}
